requirements.txt
anthropic
inspect-ai
openai
pillow
tqdm